refactor: rename with suffix Button
