+++ frontend/src/lib/components/accounts/NnsAccountsFooter.svelte
@@ -4,7 +4,7 @@
   import Footer from "$lib/components/layout/Footer.svelte";
   import IcpTransactionModal from "$lib/modals/accounts/IcpTransactionModal.svelte";
   import { nonNullish } from "@dfinity/utils";
-  import Receive from "$lib/components/accounts/Receive.svelte";
+  import Receive from "$lib/components/accounts/ReceiveButton.svelte";
 
   let modal: "NewTransaction" | undefined = undefined;
   const openNewTransaction = () => (modal = "NewTransaction");
